+++ cli/deno_dir.rs
@@ -853,58 +853,6 @@ fn save_source_code_headers(
   }
 }
 
-// https://html.spec.whatwg.org/multipage/webappapis.html#resolve-a-module-specifier
-// TODO(ry) Add tests.
-// TODO(ry) Move this to core?
-pub fn resolve_module2(
-  specifier: &str,
-  base: &str,
-) -> Result<String, url::ParseError> {
-  // 1. Apply the URL parser to specifier. If the result is not failure, return
-  //    the result.
-  // let specifier = parse_local_or_remote(specifier)?.to_string();
-  if let Ok(specifier_url) = Url::parse(specifier) {
-    return Ok(specifier_url.to_string());
-  }
-
-  // 2. If specifier does not start with the character U+002F SOLIDUS (/), the
-  //    two-character sequence U+002E FULL STOP, U+002F SOLIDUS (./), or the
-  //    three-character sequence U+002E FULL STOP, U+002E FULL STOP, U+002F
-  //    SOLIDUS (../), return failure.
-  if !specifier.starts_with("/")
-    && !specifier.starts_with("./")
-    && !specifier.starts_with("../")
-  {
-    // TODO(ry) This is (probably) not the correct error to return here.
-    return Err(url::ParseError::RelativeUrlWithCannotBeABaseBase);
-  }
-
-  // 3. Return the result of applying the URL parser to specifier with base URL
-  //    as the base URL.
-  let base_url = Url::parse(base)?;
-  let u = base_url.join(&specifier)?;
-  Ok(u.to_string())
-}
-
-/// Takes a string representing a path or URL to a module, but of the type
-/// passed through the command-line interface for the main module. This is
-/// slightly different than specifiers used in import statements: "foo.js" for
-/// example is allowed here, whereas in import statements a leading "./" is
-/// required ("./foo.js"). This function is aware of the current working
-/// directory and returns an absolute URL.
-pub fn root_specifier_to_url(
-  root_specifier: &str,
-) -> Result<Url, url::ParseError> {
-  let maybe_url = Url::parse(root_specifier);
-  if let Ok(url) = maybe_url {
-    Ok(url)
-  } else {
-    let cwd = std::env::current_dir().unwrap();
-    let base = Url::from_directory_path(cwd).unwrap();
-    base.join(root_specifier)
-  }
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
